feat(web/gui): add web viewer, ROI extraction & GUI refactor
Adds a Flask-based DeepZoom viewer and slide ROI extraction, refactors desktop GUI and inference plumbing, and modernizes training utilities.

- Introduces a web tile server and app launcher to serve slides with AI overlays for remote inspection.
- Implements ROI detection and inference dispatch (local PyTorch or remote TF serving) to provide contour overlays.
- Refactors desktop GUI and model-loading to share inference utilities and improve model discovery/fallbacks.
- Enhances PyTorch training pipeline with MixUp augmentation, Lovasz+CE combined loss, configurable AMP/scheduler, and safer checkpoint naming.
- Improves smooth-window prediction and TTA options; tightens configuration handling and adds shared inference helpers.
- Adds docs and tests to cover new loss/mixup behavior.

These changes enable a unified web/desktop workflow, easier model management, and more robust training/inference paths for slide analysis.

+++ clogic/graphics.py
@@ -41,7 +41,7 @@ def get_corrected_size(height, width, factor=256):
 
 
 def process_dense_pathology_map(pathology_map, threshold: float = 0.5):
-    markup = np.zeros((pathology_map[..., 0].shape + (4,)))
+    markup = np.zeros((pathology_map[..., 0].shape + (4,)), dtype=np.uint8)
     red_marks = markup[..., 2].copy()
     stats, texts_probs, texts_labels = {}, [], []
 
@@ -77,7 +77,7 @@ def process_dense_pathology_map(pathology_map, threshold: float = 0.5):
 
 
 def process_sparse_pathology_map(pathology_map):
-    markup = np.zeros((pathology_map.shape + (4,)))
+    markup = np.zeros((pathology_map.shape + (4,)), dtype=np.uint8)
     stats = {
         'Total objects': 0,
         'Groups': 0,
